Refactor gp server as module
