+++ src/gp_server/app/path_noise_attrs.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass, field
 from typing import List, Set, Dict, Tuple, Optional
-import app.noise_exposures as noise_exps
+import gp_server.app.noise_exposures as noise_exps
 
 
 @dataclass
