+++ src/gp_server/tests/conftest.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 import pytest
 import json
 import time
-from utils.geometry import project_geom
+from gp_server.utils.geometry import project_geom
 from shapely.geometry import LineString
 
 
@@ -13,14 +13,14 @@ __aq_sensitivities = [ 5, 15, 30 ]
 
 @pytest.fixture(scope='module')
 def initial_client():
-    patch_env_test_mode = patch('env.test_mode', True)
-    patch_env_graph_file = patch('env.graph_file', r'graphs/kumpula.graphml')
+    patch_env_test_mode = patch('gp_server.env.test_mode', True)
+    patch_env_graph_file = patch('gp_server.env.graph_file', r'graphs/kumpula.graphml')
     
-    patch_noise_sens = patch('app.noise_exposures.get_noise_sensitivities', return_value=__noise_sensitivities)
-    patch_aq_sens = patch('app.aq_exposures.get_aq_sensitivities', return_value=__aq_sensitivities)
+    patch_noise_sens = patch('gp_server.app.noise_exposures.get_noise_sensitivities', return_value=__noise_sensitivities)
+    patch_aq_sens = patch('gp_server.app.aq_exposures.get_aq_sensitivities', return_value=__aq_sensitivities)
     
     with patch_env_test_mode, patch_env_graph_file, patch_noise_sens, patch_aq_sens:
-        from green_paths_app import app
+        from gp_server_main import app
         with app.test_client() as gp_client:
             yield gp_client
