+++ src/gp_server/tests/test_api_quiet_paths.py
@@ -1,7 +1,7 @@
 from typing import Dict, Union, Tuple, Callable
 from shapely.geometry import LineString
-from utils.geometry import project_geom
-from app.constants import cost_prefix_dict, TravelMode, RoutingMode
+from gp_server.utils.geometry import project_geom
+from gp_server.app.constants import cost_prefix_dict, TravelMode, RoutingMode
 import json
 import pytest
